Prompt for password on sendtx
